feature(data-model) create vulnerability module
