+++ opencti-platform/opencti-graphql/src/database/stix-representative.ts
@@ -17,7 +17,6 @@ import {
   ENTITY_TYPE_INTRUSION_SET,
   ENTITY_TYPE_MALWARE,
   ENTITY_TYPE_TOOL,
-  ENTITY_TYPE_VULNERABILITY,
   isStixDomainObjectIdentity,
   isStixDomainObjectLocation,
   isStixDomainObjectThreatActor,
@@ -75,6 +74,7 @@ import { isStixCoreRelationship } from '../schema/stixCoreRelationship';
 import { RELATION_GRANTED_TO, RELATION_OBJECT_MARKING } from '../schema/stixRefRelationship';
 import type { BasicStoreCommon } from '../types/store';
 import { isStixSightingRelationship } from '../schema/stixSightingRelationship';
+import { ENTITY_TYPE_VULNERABILITY } from '../modules/vulnerability/vulnerability-types';
 
 export const extractStixRepresentative = (
   stix: S.StixObject,
