+++ opencti-platform/opencti-graphql/src/graphql/schema.js
@@ -46,7 +46,6 @@ import infrastructureResolvers from '../resolvers/infrastructure';
 import campaignResolvers from '../resolvers/campaign';
 import malwareResolvers from '../resolvers/malware';
 import toolResolvers from '../resolvers/tool';
-import vulnerabilityResolvers from '../resolvers/vulnerability';
 import reportResolvers from '../resolvers/report';
 import containerResolvers from '../resolvers/container';
 import noteResolvers from '../resolvers/note';
@@ -238,7 +237,6 @@ const schemaResolvers = [
   malwareResolvers,
   threatActorResolvers,
   toolResolvers,
-  vulnerabilityResolvers,
   incidentResolvers,
   // STIX CYBER OBSERVABLE ENTITIES
   stixCyberObservableResolvers,
