+++ opencti-platform/opencti-graphql/src/modules/vulnerability/vulnerability-converter.ts
@@ -1,0 +1,138 @@
+import { buildStixDomain } from '../../database/stix-2-1-converter';
+import { buildStixDomain as buildStixDomain_2_0 } from '../../database/stix-2-0-converter';
+import { assertType, cleanObject } from '../../database/stix-converter-utils';
+import { STIX_EXT_OCTI } from '../../types/stix-2-1-extensions';
+import type { StoreEntity } from '../../types/store';
+import type * as SDO from '../../types/stix-2-1-sdo';
+import type * as SDO_2_0 from '../../types/stix-2-0-sdo';
+import { ENTITY_TYPE_VULNERABILITY } from './vulnerability-types';
+
+const convertVulnerabilityToStix = (instance: StoreEntity): SDO.StixVulnerability => {
+  assertType(ENTITY_TYPE_VULNERABILITY, instance.entity_type);
+  const vulnerability = buildStixDomain(instance);
+  return {
+    ...vulnerability,
+    name: instance.name,
+    description: instance.description,
+    extensions: {
+      [STIX_EXT_OCTI]: cleanObject({
+        ...vulnerability.extensions[STIX_EXT_OCTI],
+        // CVSS3
+        cvss_vector: instance.x_opencti_cvss_vector_string,
+        cvss_base_score: instance.x_opencti_cvss_base_score,
+        cvss_base_severity: instance.x_opencti_cvss_base_severity,
+        cvss_attack_vector: instance.x_opencti_cvss_attack_vector,
+        cvss_attack_complexity: instance.x_opencti_cvss_attack_complexity,
+        cvss_privileges_required: instance.x_opencti_cvss_privileges_required,
+        cvss_user_interaction: instance.x_opencti_cvss_user_interaction,
+        cvss_scope: instance.x_opencti_cvss_scope,
+        cvss_confidentiality_impact: instance.x_opencti_cvss_confidentiality_impact,
+        cvss_integrity_impact: instance.x_opencti_cvss_integrity_impact,
+        cvss_availability_impact: instance.x_opencti_cvss_availability_impact,
+        cvss_exploit_code_maturity: instance.x_opencti_cvss_exploit_code_maturity,
+        cvss_remediation_level: instance.x_opencti_cvss_remediation_level,
+        cvss_report_confidence: instance.x_opencti_cvss_report_confidence,
+        cvss_temporal_score: instance.x_opencti_cvss_temporal_score,
+        // CVSS2
+        cvss_v2_vector: instance.x_opencti_cvss_v2_vector_string,
+        cvss_v2_base_score: instance.x_opencti_cvss_v2_base_score,
+        cvss_v2_access_vector: instance.x_opencti_cvss_v2_access_vector,
+        cvss_v2_access_complexity: instance.x_opencti_cvss_v2_access_complexity,
+        cvss_v2_authentication: instance.x_opencti_cvss_v2_authentication,
+        cvss_v2_confidentiality_impact: instance.x_opencti_cvss_v2_confidentiality_impact,
+        cvss_v2_integrity_impact: instance.x_opencti_cvss_v2_integrity_impact,
+        cvss_v2_availability_impact: instance.x_opencti_cvss_v2_availability_impact,
+        cvss_v2_exploitability: instance.x_opencti_cvss_v2_exploitability,
+        cvss_v2_remediation_level: instance.x_opencti_cvss_v2_remediation_level,
+        cvss_v2_report_confidence: instance.x_opencti_cvss_v2_report_confidence,
+        cvss_v2_temporal_score: instance.x_opencti_cvss_v2_temporal_score,
+        // CVSS4
+        cvss_v4_vector: instance.x_opencti_cvss_v4_vector_string,
+        cvss_v4_base_score: instance.x_opencti_cvss_v4_base_score,
+        cvss_v4_base_severity: instance.x_opencti_cvss_v4_base_severity,
+        cvss_v4_attack_vector: instance.x_opencti_cvss_v4_attack_vector,
+        cvss_v4_attack_complexity: instance.x_opencti_cvss_v4_attack_complexity,
+        cvss_v4_attack_requirements: instance.x_opencti_cvss_v4_attack_requirements,
+        cvss_v4_privileges_required: instance.x_opencti_cvss_v4_privileges_required,
+        cvss_v4_user_interaction: instance.x_opencti_cvss_v4_user_interaction,
+        cvss_v4_confidentiality_impact_v: instance.x_opencti_cvss_v4_confidentiality_impact_v,
+        cvss_v4_confidentiality_impact_s: instance.x_opencti_cvss_v4_confidentiality_impact_s,
+        cvss_v4_integrity_impact_v: instance.x_opencti_cvss_v4_integrity_impact_v,
+        cvss_v4_integrity_impact_s: instance.x_opencti_cvss_v4_integrity_impact_s,
+        cvss_v4_availability_impact_v: instance.x_opencti_cvss_v4_availability_impact_v,
+        cvss_v4_availability_impact_s: instance.x_opencti_cvss_v4_availability_impact_s,
+        cvss_v4_exploit_maturity: instance.x_opencti_cvss_v4_exploit_maturity,
+        // Others
+        cwe: instance.x_opencti_cwe,
+        cisa_kev: instance.x_opencti_cisa_kev,
+        epss_score: instance.x_opencti_epss_score,
+        epss_percentile: instance.x_opencti_epss_percentile,
+        score: instance.x_opencti_score,
+        first_seen_active: instance.x_opencti_first_seen_active,
+      }),
+    },
+  };
+};
+
+export const convertVulnerabilityToStix_2_0 = (instance: StoreEntity): SDO_2_0.StixVulnerability => {
+  assertType(ENTITY_TYPE_VULNERABILITY, instance.entity_type);
+  const vulnerability = buildStixDomain_2_0(instance);
+  return {
+    ...vulnerability,
+    name: instance.name,
+    description: instance.description,
+    x_opencti_cisa_kev: instance.x_opencti_cisa_kev,
+    x_opencti_first_seen_active: instance.x_opencti_first_seen_active,
+    // CVSS3
+    x_opencti_cvss_vector_string: instance.x_opencti_cvss_vector_string,
+    x_opencti_cvss_base_score: instance.x_opencti_cvss_base_score,
+    x_opencti_cvss_base_severity: instance.x_opencti_cvss_base_severity,
+    x_opencti_cvss_attack_vector: instance.x_opencti_cvss_attack_vector,
+    x_opencti_cvss_attack_complexity: instance.x_opencti_cvss_attack_complexity,
+    x_opencti_cvss_privileges_required: instance.x_opencti_cvss_privileges_required,
+    x_opencti_cvss_user_interaction: instance.x_opencti_cvss_user_interaction,
+    x_opencti_cvss_scope: instance.x_opencti_cvss_scope,
+    x_opencti_cvss_confidentiality_impact: instance.x_opencti_cvss_confidentiality_impact,
+    x_opencti_cvss_integrity_impact: instance.x_opencti_cvss_integrity_impact,
+    x_opencti_cvss_availability_impact: instance.x_opencti_cvss_availability_impact,
+    x_opencti_cvss_exploit_code_maturity: instance.x_opencti_cvss_exploit_code_maturity,
+    x_opencti_cvss_remediation_level: instance.x_opencti_cvss_remediation_level,
+    x_opencti_cvss_report_confidence: instance.x_opencti_cvss_report_confidence,
+    x_opencti_cvss_temporal_score: instance.x_opencti_cvss_temporal_score,
+    // CVSS2
+    x_opencti_cvss_v2_vector_string: instance.x_opencti_cvss_v2_vector_string,
+    x_opencti_cvss_v2_base_score: instance.x_opencti_cvss_v2_base_score,
+    x_opencti_cvss_v2_access_vector: instance.x_opencti_cvss_v2_access_vector,
+    x_opencti_cvss_v2_access_complexity: instance.x_opencti_cvss_v2_access_complexity,
+    x_opencti_cvss_v2_authentication: instance.x_opencti_cvss_v2_authentication,
+    x_opencti_cvss_v2_confidentiality_impact: instance.x_opencti_cvss_v2_confidentiality_impact,
+    x_opencti_cvss_v2_integrity_impact: instance.x_opencti_cvss_v2_integrity_impact,
+    x_opencti_cvss_v2_availability_impact: instance.x_opencti_cvss_v2_availability_impact,
+    x_opencti_cvss_v2_exploitability: instance.x_opencti_cvss_v2_exploitability,
+    x_opencti_cvss_v2_remediation_level: instance.x_opencti_cvss_v2_remediation_level,
+    x_opencti_cvss_v2_report_confidence: instance.x_opencti_cvss_v2_report_confidence,
+    x_opencti_cvss_v2_temporal_score: instance.x_opencti_cvss_v2_temporal_score,
+    // CVSS4
+    x_opencti_cvss_v4_vector_string: instance.x_opencti_cvss_v4_vector_string,
+    x_opencti_cvss_v4_base_score: instance.x_opencti_cvss_v4_base_score,
+    x_opencti_cvss_v4_base_severity: instance.x_opencti_cvss_v4_base_severity,
+    x_opencti_cvss_v4_attack_vector: instance.x_opencti_cvss_v4_attack_vector,
+    x_opencti_cvss_v4_attack_complexity: instance.x_opencti_cvss_v4_attack_complexity,
+    x_opencti_cvss_v4_attack_requirements: instance.x_opencti_cvss_v4_attack_requirements,
+    x_opencti_cvss_v4_privileges_required: instance.x_opencti_cvss_v4_privileges_required,
+    x_opencti_cvss_v4_user_interaction: instance.x_opencti_cvss_v4_user_interaction,
+    x_opencti_cvss_v4_confidentiality_impact_v: instance.x_opencti_cvss_v4_confidentiality_impact_v,
+    x_opencti_cvss_v4_confidentiality_impact_s: instance.x_opencti_cvss_v4_confidentiality_impact_s,
+    x_opencti_cvss_v4_integrity_impact_v: instance.x_opencti_cvss_v4_integrity_impact_v,
+    x_opencti_cvss_v4_integrity_impact_s: instance.x_opencti_cvss_v4_integrity_impact_s,
+    x_opencti_cvss_v4_availability_impact_v: instance.x_opencti_cvss_v4_availability_impact_v,
+    x_opencti_cvss_v4_availability_impact_s: instance.x_opencti_cvss_v4_availability_impact_s,
+    x_opencti_cvss_v4_exploit_maturity: instance.x_opencti_cvss_v4_exploit_maturity,
+    // Others
+    x_opencti_score: instance.x_opencti_score,
+    x_opencti_epss_score: instance.x_opencti_epss_score,
+    x_opencti_epss_percentile: instance.x_opencti_epss_percentile,
+  };
+};
+
+export default convertVulnerabilityToStix;
