+++ opencti-platform/opencti-graphql/src/modules/vulnerability/vulnerability-domain.ts
@@ -1,27 +1,30 @@
-import { createEntity } from '../database/middleware';
-import { pageEntitiesConnection, pageRegardingEntitiesConnection, storeLoadById } from '../database/middleware-loader';
-import { BUS_TOPICS } from '../config/conf';
-import { notify } from '../database/redis';
-import { ENTITY_TYPE_VULNERABILITY } from '../schema/stixDomainObject';
-import { ABSTRACT_STIX_DOMAIN_OBJECT } from '../schema/general';
-import { RELATION_HAS } from '../schema/stixCoreRelationship';
-import { ENTITY_SOFTWARE } from '../schema/stixCyberObservable';
-import { isValidCvssVector, parseCvssVector, updateCvssVector } from '../utils/vulnerabilities';
-import { FunctionalError } from '../config/errors';
-import { isNotEmptyField } from '../database/utils';
+import { createEntity } from '../../database/middleware';
+import { pageEntitiesConnection, pageRegardingEntitiesConnection, storeLoadById } from '../../database/middleware-loader';
+import { BUS_TOPICS } from '../../config/conf';
+import { notify } from '../../database/redis';
+import { ABSTRACT_STIX_DOMAIN_OBJECT } from '../../schema/general';
+import { RELATION_HAS } from '../../schema/stixCoreRelationship';
+import { ENTITY_SOFTWARE } from '../../schema/stixCyberObservable';
+import { isValidCvssVector, parseCvssVector, updateCvssVector } from '../../utils/vulnerabilities';
+import { FunctionalError } from '../../config/errors';
+import { isNotEmptyField } from '../../database/utils';
+import { CVSS_SEVERITY_VALUES, ENTITY_TYPE_VULNERABILITY } from './vulnerability-types';
+import type { AuthContext, AuthUser } from '../../types/user';
+import type { EntityOptions } from '../../database/middleware-loader';
+import type { BasicStoreEntityVulnerability } from './vulnerability-types';
 
-export const CVSS_SEVERITY_VALUES = ['CRITICAL', 'HIGH', 'MEDIUM', 'LOW', 'Unknown'];
+export { CVSS_SEVERITY_VALUES };
 
-export const findById = (context, user, vulnerabilityId) => {
-  return storeLoadById(context, user, vulnerabilityId, ENTITY_TYPE_VULNERABILITY);
+export const findById = (context: AuthContext, user: AuthUser, vulnerabilityId: string) => {
+  return storeLoadById<BasicStoreEntityVulnerability>(context, user, vulnerabilityId, ENTITY_TYPE_VULNERABILITY);
 };
 
-export const findVulnerabilityPaginated = (context, user, args) => {
-  return pageEntitiesConnection(context, user, [ENTITY_TYPE_VULNERABILITY], args);
+export const findVulnerabilityPaginated = (context: AuthContext, user: AuthUser, args: EntityOptions<BasicStoreEntityVulnerability>) => {
+  return pageEntitiesConnection<BasicStoreEntityVulnerability>(context, user, [ENTITY_TYPE_VULNERABILITY], args);
 };
 
-export const addVulnerability = async (context, user, vulnerability) => {
-  let finalVulnerability = vulnerability;
+export const addVulnerability = async (context: AuthContext, user: AuthUser, vulnerability: any) => {
+  let finalVulnerability: any = vulnerability;
 
   // CVSS 2
   if (isNotEmptyField(vulnerability.x_opencti_cvss_v2_vector_string)) {
@@ -91,6 +94,6 @@ export const addVulnerability = async (context, user, vulnerability) => {
   return notify(BUS_TOPICS[ABSTRACT_STIX_DOMAIN_OBJECT].ADDED_TOPIC, created, user);
 };
 
-export const softwarePaginated = async (context, user, vulnerabilityId, opts) => {
+export const softwarePaginated = async (context: AuthContext, user: AuthUser, vulnerabilityId: string, opts: any) => {
   return pageRegardingEntitiesConnection(context, user, vulnerabilityId, opts.relationshipType ?? RELATION_HAS, ENTITY_SOFTWARE, true, opts);
 };
