+++ opencti-platform/opencti-graphql/src/modules/vulnerability/vulnerability-graphql.ts
@@ -1,0 +1,8 @@
+import { registerGraphqlSchema } from '../../graphql/schema';
+import vulnerabilityTypeDefs from './vulnerability.graphql';
+import vulnerabilityResolvers from './vulnerability-resolver';
+
+registerGraphqlSchema({
+  schema: vulnerabilityTypeDefs,
+  resolver: vulnerabilityResolvers,
+});
