+++ opencti-platform/opencti-graphql/src/modules/vulnerability/vulnerability-resolver.ts
@@ -1,0 +1,36 @@
+import type { AuthContext, AuthUser } from '../../types/user';
+import { addVulnerability, findById, findVulnerabilityPaginated, softwarePaginated } from './vulnerability-domain';
+import {
+  stixDomainObjectAddRelation,
+  stixDomainObjectCleanContext,
+  stixDomainObjectDelete,
+  stixDomainObjectDeleteRelation,
+  stixDomainObjectEditContext,
+  stixDomainObjectEditField,
+} from '../../domain/stixDomainObject';
+import { ENTITY_TYPE_VULNERABILITY } from './vulnerability-types';
+
+type VulnerabilityContext = AuthContext & { user: AuthUser };
+
+const vulnerabilityResolvers = {
+  Query: {
+    vulnerability: (_: unknown, { id }: { id: string }, context: VulnerabilityContext) => findById(context, context.user, id),
+    vulnerabilities: (_: unknown, args: any, context: VulnerabilityContext) => findVulnerabilityPaginated(context, context.user, args),
+  },
+  Vulnerability: {
+    softwares: (vulnerability: { id: string }, args: any, context: VulnerabilityContext) => softwarePaginated(context, context.user, vulnerability.id, args),
+  },
+  Mutation: {
+    vulnerabilityEdit: (_: unknown, { id }: { id: string }, context: VulnerabilityContext) => ({
+      delete: () => stixDomainObjectDelete(context, context.user, id, ENTITY_TYPE_VULNERABILITY),
+      fieldPatch: ({ input, commitMessage, references }: any) => stixDomainObjectEditField(context, context.user, id, input, { commitMessage, references }),
+      contextPatch: ({ input }: any) => stixDomainObjectEditContext(context, context.user, id, input),
+      contextClean: () => stixDomainObjectCleanContext(context, context.user, id),
+      relationAdd: ({ input }: any) => stixDomainObjectAddRelation(context, context.user, id, input),
+      relationDelete: ({ toId, relationship_type: relationshipType }: any) => stixDomainObjectDeleteRelation(context, context.user, id, toId, relationshipType),
+    }),
+    vulnerabilityAdd: (_: unknown, { input }: any, context: VulnerabilityContext) => addVulnerability(context, context.user, input),
+  },
+};
+
+export default vulnerabilityResolvers;
