+++ opencti-platform/opencti-graphql/src/modules/vulnerability/vulnerability-types.ts
@@ -1,0 +1,17 @@
+import type { BasicStoreEntity, StoreEntity } from '../../types/store';
+
+export const ENTITY_TYPE_VULNERABILITY = 'Vulnerability';
+
+export const CVSS_SEVERITY_VALUES = ['CRITICAL', 'HIGH', 'MEDIUM', 'LOW', 'Unknown'];
+
+// region Database types
+export interface BasicStoreEntityVulnerability extends BasicStoreEntity {
+  name: string;
+  description: string;
+}
+
+export interface StoreEntityVulnerability extends StoreEntity {
+  name: string;
+  description: string;
+}
+// endregion
