--- opencti-platform/opencti-graphql/src/resolvers/vulnerability.js
@@ -1,33 +1,0 @@
-import { addVulnerability, findVulnerabilityPaginated, findById, softwarePaginated } from '../domain/vulnerability';
-import {
-  stixDomainObjectAddRelation,
-  stixDomainObjectCleanContext,
-  stixDomainObjectDeleteRelation,
-  stixDomainObjectDelete,
-  stixDomainObjectEditContext,
-  stixDomainObjectEditField,
-} from '../domain/stixDomainObject';
-import { ENTITY_TYPE_VULNERABILITY } from '../schema/stixDomainObject';
-
-const vulnerabilityResolvers = {
-  Query: {
-    vulnerability: (_, { id }, context) => findById(context, context.user, id),
-    vulnerabilities: (_, args, context) => findVulnerabilityPaginated(context, context.user, args),
-  },
-  Vulnerability: {
-    softwares: (vulnerability, args, context) => softwarePaginated(context, context.user, vulnerability.id, args),
-  },
-  Mutation: {
-    vulnerabilityEdit: (_, { id }, context) => ({
-      delete: () => stixDomainObjectDelete(context, context.user, id, ENTITY_TYPE_VULNERABILITY),
-      fieldPatch: ({ input, commitMessage, references }) => stixDomainObjectEditField(context, context.user, id, input, { commitMessage, references }),
-      contextPatch: ({ input }) => stixDomainObjectEditContext(context, context.user, id, input),
-      contextClean: () => stixDomainObjectCleanContext(context, context.user, id),
-      relationAdd: ({ input }) => stixDomainObjectAddRelation(context, context.user, id, input),
-      relationDelete: ({ toId, relationship_type: relationshipType }) => stixDomainObjectDeleteRelation(context, context.user, id, toId, relationshipType),
-    }),
-    vulnerabilityAdd: (_, { input }, context) => addVulnerability(context, context.user, input),
-  },
-};
-
-export default vulnerabilityResolvers;
