+++ opencti-platform/opencti-graphql/src/rules/vulnerability-software-to-infra/VulnerabilitySoftwareToInfraDefinition.ts
@@ -1,7 +1,8 @@
 import { RELATION_CONSISTS_OF, RELATION_HAS } from '../../schema/stixCoreRelationship';
 import { ENTITY_SOFTWARE } from '../../schema/stixCyberObservable';
-import { ENTITY_TYPE_INFRASTRUCTURE, ENTITY_TYPE_VULNERABILITY } from '../../schema/stixDomainObject';
+import { ENTITY_TYPE_INFRASTRUCTURE } from '../../schema/stixDomainObject';
 import type { RuleDefinition } from '../../types/rules';
+import { ENTITY_TYPE_VULNERABILITY } from '../../modules/vulnerability/vulnerability-types';
 
 const id = 'vulnerability_software_to_infra';
 const name = 'Software Vulnerability propagation to Infrastructure';
