+++ opencti-platform/opencti-graphql/src/rules/vulnerability-software-to-infra/VulnerabilitySoftwareToInfraRule.ts
@@ -1,7 +1,8 @@
 import { RELATION_CONSISTS_OF, RELATION_HAS } from '../../schema/stixCoreRelationship';
-import { ENTITY_TYPE_INFRASTRUCTURE, ENTITY_TYPE_VULNERABILITY } from '../../schema/stixDomainObject';
+import { ENTITY_TYPE_INFRASTRUCTURE } from '../../schema/stixDomainObject';
 import buildRelationToRelationRule from '../relationToRelationBuilder';
 import def from './VulnerabilitySoftwareToInfraDefinition';
+import { ENTITY_TYPE_VULNERABILITY } from '../../modules/vulnerability/vulnerability-types';
 
 /**
  * Rule: If Infrastructure A consists of Software B, If Software B has Vulnerability C,
