+++ opencti-platform/opencti-graphql/src/rules/vulnerability-software-to-system/VulnerabilitySoftwareToSystemDefinition.ts
@@ -1,7 +1,8 @@
 import { RELATION_HAS, RELATION_USES } from '../../schema/stixCoreRelationship';
 import { ENTITY_SOFTWARE } from '../../schema/stixCyberObservable';
-import { ENTITY_TYPE_IDENTITY_SYSTEM, ENTITY_TYPE_VULNERABILITY } from '../../schema/stixDomainObject';
+import { ENTITY_TYPE_IDENTITY_SYSTEM } from '../../schema/stixDomainObject';
 import type { RuleDefinition } from '../../types/rules';
+import { ENTITY_TYPE_VULNERABILITY } from '../../modules/vulnerability/vulnerability-types';
 
 const id = 'vulnerability_software_to_system';
 const name = 'Vulnerability propagation via software usage';
