+++ opencti-platform/opencti-graphql/src/rules/vulnerability-software-to-system/VulnerabilitySoftwareToSystemRule.ts
@@ -1,7 +1,8 @@
 import { RELATION_HAS, RELATION_USES } from '../../schema/stixCoreRelationship';
-import { ENTITY_TYPE_IDENTITY_SYSTEM, ENTITY_TYPE_VULNERABILITY } from '../../schema/stixDomainObject';
+import { ENTITY_TYPE_IDENTITY_SYSTEM } from '../../schema/stixDomainObject';
 import buildRelationToRelationRule from '../relationToRelationBuilder';
 import def from './VulnerabilitySoftwareToSystemDefinition';
+import { ENTITY_TYPE_VULNERABILITY } from '../../modules/vulnerability/vulnerability-types';
 
 /**
  * Rule: If System A uses Software B, If Software B has Vulnerability C,
