+++ opencti-platform/opencti-graphql/tests/03-integration/01-database/filterKeysSchema-test.ts
@@ -13,7 +13,7 @@ import {
   INPUT_LABELS,
 } from '../../../src/schema/general';
 import { ENTITY_TYPE_LABEL } from '../../../src/schema/stixMetaObject';
-import { ENTITY_TYPE_ATTACK_PATTERN, ENTITY_TYPE_CONTAINER_REPORT, ENTITY_TYPE_MALWARE, ENTITY_TYPE_VULNERABILITY } from '../../../src/schema/stixDomainObject';
+import { ENTITY_TYPE_ATTACK_PATTERN, ENTITY_TYPE_CONTAINER_REPORT, ENTITY_TYPE_MALWARE } from '../../../src/schema/stixDomainObject';
 import { ENTITY_TYPE_NOTIFICATION, ENTITY_TYPE_TRIGGER } from '../../../src/modules/notification/notification-types';
 import {
   ENTITY_HASHED_OBSERVABLE_ARTIFACT,
@@ -42,6 +42,7 @@ import { ENTITY_TYPE_STATUS_TEMPLATE } from '../../../src/schema/internalObject'
 import { ENTITY_TYPE_DRAFT_WORKSPACE } from '../../../src/modules/draftWorkspace/draftWorkspace-types';
 import { WORKFLOW_INSTANCE_STATUS_FILTER } from '../../../src/utils/filtering/filtering-constants';
 import { ENTITY_TYPE_CONTAINER_GROUPING } from '../../../src/modules/grouping/grouping-types';
+import { ENTITY_TYPE_VULNERABILITY } from '../../../src/modules/vulnerability/vulnerability-types';
 
 describe('Filter keys schema generation testing', async () => {
   const filterKeysSchemaArray = await generateFilterKeysSchema();
